Update bots to used shared library
