+++ examples/medplum-chart-demo/esbuild-script.mjs
@@ -1,0 +1,41 @@
+/* global console */
+/* global process */
+/* eslint no-console: "off" */
+/*eslint no-process-exit: "off"*/
+
+import esbuild from 'esbuild';
+import { glob } from 'glob';
+import botLayer from '@medplum/bot-layer/package.json' assert { type: 'json' };
+
+// Find all TypeScript files in your source directory
+const entryPoints = glob.sync('./src/**/*.ts').filter((file) => !file.endsWith('test.ts'));
+
+const botLayerDeps = Object.keys(botLayer.dependencies);
+
+// Define the esbuild options
+const esbuildOptions = {
+  entryPoints: entryPoints,
+  bundle: true, // Bundle imported functions
+  outdir: './dist', // Output directory for compiled files
+  platform: 'node', // or 'node', depending on your target platform
+  loader: {
+    '.ts': 'ts', // Load TypeScript files
+  },
+  resolveExtensions: ['.ts'],
+  external: botLayerDeps,
+  format: 'cjs', // Set output format as ECMAScript modules
+  target: 'es2020', // Set the target ECMAScript version
+  tsconfig: 'tsconfig.json',
+  footer: { js: 'Object.assign(exports, module.exports);' }, // Required for VM Context Bots
+};
+
+// Build using esbuild
+esbuild
+  .build(esbuildOptions)
+  .then(() => {
+    console.log('Build completed successfully!');
+  })
+  .catch((error) => {
+    console.error('Build failed:', JSON.stringify(error, null, 2));
+    process.exit(1);
+  });
