+++ examples/medplum-chart-demo/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "build": "tsc && vite build",
-    "build:bots": "npm run clean && npm run lint && tsc --project tsconfig-bots.json && node --loader ts-node/esm src/scripts/deploy-bots.ts",
+    "build:bots": "npm run clean && npm run lint && tsc --project tsconfig-bots.json && node --loader esbuild-script.mjs ts-node/esm src/scripts/deploy-bots.ts",
     "clean": "rimraf dist",
     "lint": "eslint src/",
     "dev": "vite",
@@ -37,6 +37,7 @@
     "@types/react-dom": "18.3.0",
     "@vitejs/plugin-react": "4.3.0",
     "chart.js": "^4.4.2",
+    "esbuild": "0.21.2",
     "postcss": "8.4.38",
     "postcss-preset-mantine": "1.15.0",
     "react": "18.2.0",
