+++ examples/medplum-chart-demo/src/bots/core/test-data/general-encounter-test-data.ts
@@ -10,6 +10,7 @@ export const fullResponse: QuestionnaireResponse = {
   resourceType: 'QuestionnaireResponse',
   status: 'completed',
   item: [
+    { linkId: 'date', answer: [{ valueDateTime: '2024-02-14T11:18:05.446Z' }] },
     {
       linkId: 'reason-for-visit',
       answer: [{ valueCoding: { code: '112233', system: 'http://hl7.org/fhir/sid/icd-10', display: 'DIAG-4' } }],
@@ -32,6 +33,7 @@ export const fullResponseNoProblemList: QuestionnaireResponse = {
   resourceType: 'QuestionnaireResponse',
   status: 'completed',
   item: [
+    { linkId: 'date', answer: [{ valueDateTime: '2024-02-14T11:18:05.446Z' }] },
     {
       linkId: 'reason-for-visit',
       answer: [{ valueCoding: { code: '112233', system: 'http://hl7.org/fhir/sid/icd-10', display: 'DIAG-4' } }],
@@ -45,7 +47,7 @@ export const fullResponseNoProblemList: QuestionnaireResponse = {
     { linkId: 'mood-swings', answer: [{ valueBoolean: true }] },
     { linkId: 'vaginal-dryness', answer: [{ valueBoolean: true }] },
     { linkId: 'sleep-disturbance', answer: [{ valueBoolean: true }] },
-    { linkId: 'self-reported-history', answer: [{ valueString: 'Blood clots' }] },
+    { linkId: 'self-reported-history', answer: [{ valueString: 'Irregular bleeding' }] },
     { linkId: 'assessment', answer: [{ valueString: 'Everything looks bad' }] },
   ],
 };
@@ -54,6 +56,7 @@ export const noReasonForVisit: QuestionnaireResponse = {
   resourceType: 'QuestionnaireResponse',
   status: 'completed',
   item: [
+    { linkId: 'date', answer: [{ valueDateTime: '2024-02-14T11:18:05.446Z' }] },
     { linkId: 'systolic', answer: [{ valueInteger: 180 }] },
     { linkId: 'diastolic', answer: [{ valueInteger: 180 }] },
     { linkId: 'height', answer: [{ valueQuantity: { value: 180, unit: 'cm' } }] },
@@ -62,7 +65,7 @@ export const noReasonForVisit: QuestionnaireResponse = {
     { linkId: 'mood-swings', answer: [{ valueBoolean: true }] },
     { linkId: 'vaginal-dryness', answer: [{ valueBoolean: true }] },
     { linkId: 'sleep-disturbance', answer: [{ valueBoolean: true }] },
-    { linkId: 'self-reported-history', answer: [{ valueString: 'Blood clots' }] },
+    { linkId: 'self-reported-history', answer: [{ valueString: 'Breast cancer' }] },
     { linkId: 'assessment', answer: [{ valueString: 'Everything looks bad' }] },
   ],
 };
@@ -71,6 +74,7 @@ export const onlyCondition: QuestionnaireResponse = {
   resourceType: 'QuestionnaireResponse',
   status: 'completed',
   item: [
+    { linkId: 'date', answer: [{ valueDateTime: '2024-02-14T11:18:05.446Z' }] },
     {
       linkId: 'reason-for-visit',
       answer: [{ valueCoding: { code: '112233', system: 'http://hl7.org/fhir/sid/icd-10', display: 'DIAG-4' } }],
@@ -83,6 +87,7 @@ export const noSelfReportedHistory: QuestionnaireResponse = {
   resourceType: 'QuestionnaireResponse',
   status: 'completed',
   item: [
+    { linkId: 'date', answer: [{ valueDateTime: '2024-02-14T11:18:05.446Z' }] },
     {
       linkId: 'reason-for-visit',
       answer: [{ valueCoding: { code: '112233', system: 'http://hl7.org/fhir/sid/icd-10', display: 'DIAG-4' } }],
@@ -104,6 +109,7 @@ export const oneBloodPressureMeasurement: QuestionnaireResponse = {
   resourceType: 'QuestionnaireResponse',
   status: 'completed',
   item: [
+    { linkId: 'date', answer: [{ valueDateTime: '2024-02-14T11:18:05.446Z' }] },
     {
       linkId: 'reason-for-visit',
       answer: [{ valueCoding: { code: '112233', system: 'http://hl7.org/fhir/sid/icd-10', display: 'DIAG-4' } }],
@@ -116,7 +122,73 @@ export const oneBloodPressureMeasurement: QuestionnaireResponse = {
     { linkId: 'mood-swings', answer: [{ valueBoolean: true }] },
     { linkId: 'vaginal-dryness', answer: [{ valueBoolean: true }] },
     { linkId: 'sleep-disturbance', answer: [{ valueBoolean: true }] },
-    { linkId: 'self-reported-history', answer: [{ valueString: 'Blood clots' }] },
+    { linkId: 'self-reported-history', answer: [{ valueString: 'Stroke' }] },
+    { linkId: 'assessment', answer: [{ valueString: 'Everything looks bad' }] },
+  ],
+};
+
+export const selfReportedHistoryBmi: QuestionnaireResponse = {
+  resourceType: 'QuestionnaireResponse',
+  status: 'completed',
+  item: [
+    { linkId: 'date', answer: [{ valueDateTime: '2024-02-14T11:18:05.446Z' }] },
+    {
+      linkId: 'reason-for-visit',
+      answer: [{ valueCoding: { code: '112233', system: 'http://hl7.org/fhir/sid/icd-10', display: 'DIAG-4' } }],
+    },
+    { linkId: 'problem-list', answer: [{ valueBoolean: true }] },
+    { linkId: 'diastolic', answer: [{ valueInteger: 180 }] },
+    { linkId: 'height', answer: [{ valueQuantity: { value: 180, unit: 'cm' } }] },
+    { linkId: 'weight', answer: [{ valueQuantity: { value: 150, unit: 'lb' } }] },
+    { linkId: 'hot-flashes', answer: [{ valueBoolean: true }] },
+    { linkId: 'mood-swings', answer: [{ valueBoolean: true }] },
+    { linkId: 'vaginal-dryness', answer: [{ valueBoolean: true }] },
+    { linkId: 'sleep-disturbance', answer: [{ valueBoolean: true }] },
+    { linkId: 'self-reported-history', answer: [{ valueString: 'BMI > 30' }] },
+    { linkId: 'assessment', answer: [{ valueString: 'Everything looks bad' }] },
+  ],
+};
+
+export const selfReportedHistoryBreastCancer: QuestionnaireResponse = {
+  resourceType: 'QuestionnaireResponse',
+  status: 'completed',
+  item: [
+    { linkId: 'date', answer: [{ valueDateTime: '2024-02-14T11:18:05.446Z' }] },
+    {
+      linkId: 'reason-for-visit',
+      answer: [{ valueCoding: { code: '112233', system: 'http://hl7.org/fhir/sid/icd-10', display: 'DIAG-4' } }],
+    },
+    { linkId: 'problem-list', answer: [{ valueBoolean: true }] },
+    { linkId: 'diastolic', answer: [{ valueInteger: 180 }] },
+    { linkId: 'height', answer: [{ valueQuantity: { value: 180, unit: 'cm' } }] },
+    { linkId: 'weight', answer: [{ valueQuantity: { value: 150, unit: 'lb' } }] },
+    { linkId: 'hot-flashes', answer: [{ valueBoolean: true }] },
+    { linkId: 'mood-swings', answer: [{ valueBoolean: true }] },
+    { linkId: 'vaginal-dryness', answer: [{ valueBoolean: true }] },
+    { linkId: 'sleep-disturbance', answer: [{ valueBoolean: true }] },
+    { linkId: 'self-reported-history', answer: [{ valueString: 'Breast cancer' }] },
+    { linkId: 'assessment', answer: [{ valueString: 'Everything looks bad' }] },
+  ],
+};
+
+export const selfReportedHistoryEndometrialCancer: QuestionnaireResponse = {
+  resourceType: 'QuestionnaireResponse',
+  status: 'completed',
+  item: [
+    { linkId: 'date', answer: [{ valueDateTime: '2024-02-14T11:18:05.446Z' }] },
+    {
+      linkId: 'reason-for-visit',
+      answer: [{ valueCoding: { code: '112233', system: 'http://hl7.org/fhir/sid/icd-10', display: 'DIAG-4' } }],
+    },
+    { linkId: 'problem-list', answer: [{ valueBoolean: true }] },
+    { linkId: 'diastolic', answer: [{ valueInteger: 180 }] },
+    { linkId: 'height', answer: [{ valueQuantity: { value: 180, unit: 'cm' } }] },
+    { linkId: 'weight', answer: [{ valueQuantity: { value: 150, unit: 'lb' } }] },
+    { linkId: 'hot-flashes', answer: [{ valueBoolean: true }] },
+    { linkId: 'mood-swings', answer: [{ valueBoolean: true }] },
+    { linkId: 'vaginal-dryness', answer: [{ valueBoolean: true }] },
+    { linkId: 'sleep-disturbance', answer: [{ valueBoolean: true }] },
+    { linkId: 'self-reported-history', answer: [{ valueString: 'Endometrial cancer' }] },
     { linkId: 'assessment', answer: [{ valueString: 'Everything looks bad' }] },
   ],
 };
